(client) feat: improve service-worker - prolong GraphQL API cache indefinitely, since it uses network-only strategy so user will always get latest data on online - on codebase update, reload the page to get the latest changes - add limit on caching lottie and json files
