+++ client/src/index.tsx
@@ -3,23 +3,34 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
-import reportWebVitals from "./reportWebVitals";
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://cra.link/PWA
-serviceWorkerRegistration.register().then(() => {
-  const root = ReactDOM.createRoot(
-    document.getElementById("root") as HTMLElement,
-  );
-  root.render(
-    <React.StrictMode>
-      <App />
-    </React.StrictMode>,
-  );
+serviceWorkerRegistration
+  .register({
+    onUpdate: (registration) => {
+      const waitingServiceWorker = registration.waiting;
 
-  // If you want to start measuring performance in your app, pass a function
-  // to log results (for example: reportWebVitals(console.log))
-  // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-  reportWebVitals();
-});
+      if (waitingServiceWorker) {
+        waitingServiceWorker.addEventListener("statechange", (event) => {
+          if (
+            (event.target as null | { state: string })?.state === "activated"
+          ) {
+            window.location.reload();
+          }
+        });
+        waitingServiceWorker.postMessage({ type: "SKIP_WAITING" });
+      }
+    },
+  })
+  .then(() => {
+    const root = ReactDOM.createRoot(
+      document.getElementById("root") as HTMLElement,
+    );
+    root.render(
+      <React.StrictMode>
+        <App />
+      </React.StrictMode>,
+    );
+  });
